+++ pyxel/tilemap_wrapper.py
@@ -23,11 +23,20 @@ def setup_apis(module, lib):
         def data(self):
             return self._data
 
+        @property
+        def refimg(self):
+            return lib.tilemap_refimg_getter(ctypes.c_void_p(self._c_obj))
+
+        @refimg.setter
+        def refimg(self, value):
+            return lib.tilemap_refimg_setter(ctypes.c_void_p(self._c_obj), value)
+
         def get(self, x, y):
             return lib.tilemap_get(self._c_obj, x, y)
 
-        def set(self, x, y, data, refimg):
-            pass
+        def set(self, x, y, data):
+            lib.tilemap_set1(self._c_obj, x, y, data)
+            # todo
 
         def copy(self, x, y, tm, u, v, w, h):
             lib.tilemap_copy(self._c_obj, x, y, tm, u, v, w, h)